--- src/services/auth/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './request-dto';
-export * from './response-dto';
